Move AppBar from StepPracticeFragment & StepTheoryFragment to StepFragment. Move StepQuizFeature running to StepWrapperFragment.
