+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/step/view/model/StepToolbarHost.kt
@@ -1,0 +1,5 @@
+package org.hyperskill.app.android.step.view.model
+
+interface StepToolbarHost {
+    fun render(viewState: StepToolbarViewState)
+}
