+++ androidHyperskillApp/src/main/res/layout/fragment_step_practice.xml
@@ -4,13 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <include
-        android:id="@+id/stepPracticeAppBar"
-        layout="@layout/view_step_quiz_appbar"
-        app:layout_constraintTop_toTopOf="parent"/>
+    android:layout_height="match_parent">
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/stepPracticeContainer"
@@ -18,7 +12,7 @@
         android:layout_height="0dp"
         android:fillViewport="true"
         android:scrollbars="vertical"
-        app:layout_constraintTop_toBottomOf="@+id/stepPracticeAppBar"
+        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toTopOf="@+id/stepQuizFillBlanksOptionsContainer">
 
         <androidx.appcompat.widget.LinearLayoutCompat
