+++ androidHyperskillApp/src/main/res/layout/fragment_step_theory.xml
@@ -6,16 +6,11 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <include
-        android:id="@+id/stepTheoryAppBar"
-        layout="@layout/view_step_appbar"
-        app:layout_constraintTop_toTopOf="parent"/>
-
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:fillViewport="true"
-        app:layout_constraintTop_toBottomOf="@+id/stepTheoryAppBar"
+        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent">
 
         <LinearLayout
