+++ androidHyperskillApp/src/main/res/layout/view_step_appbar.xml
@@ -1,11 +1,54 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.appbar.AppBarLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.google.android.material.appbar.AppBarLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:elevation="?appBarElevation">
 
-    <include
+    <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/stepToolbar"
-        layout="@layout/view_step_toolbar" />
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/color_background"
+        app:contentInsetStartWithNavigation="0dp"
+        app:navigationIcon="@drawable/ic_toolbar_back"
+        app:navigationIconTint="?colorControlNormal"
+        app:title=" " >
+
+        <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <TextView
+                android:id="@+id/stepQuizLimitsTextView"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textSize="18sp"
+                android:textColor="@color/color_on_surface_alpha_60"
+                android:layout_gravity="center"
+                app:drawableStartCompat="@drawable/ic_limits"
+                android:drawablePadding="4dp"
+                tools:text="10"
+                android:visibility="gone"
+                tools:visibility="visible"/>
+
+            <TextView
+                android:id="@+id/stepTheoryToolbarTitle"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginEnd="@dimen/step_toolbar_margin_end"
+                android:ellipsize="end"
+                android:maxLines="2"
+                android:textAppearance="?textAppearanceSubtitle1"
+                android:textColor="@color/material_on_surface_emphasis_high_type"
+                android:textSize="18sp"
+                tools:text="Matrix representation of systems of linear equations"
+                android:visibility="gone"/>
+
+        </FrameLayout>
+
+    </com.google.android.material.appbar.MaterialToolbar>
 
 </com.google.android.material.appbar.AppBarLayout>
